Minor changes on basicconcept

+++ semtag-core/src/main/java/org/semtag/core/Concept.java
@@ -20,5 +20,5 @@ public abstract class Concept {
         return type;
     }
 
-    public abstract float getSimilarityTo(Concept other);
+    public abstract double getSimilarityTo(Concept other);
 }
